Changed image to file
