+++ functions/routes/file.route.js
@@ -1,0 +1,8 @@
+const express = require("express");
+const fileController = require("../controllers/file.controller");
+
+const router = express.Router();
+
+router.get("/get", fileController.getFiles);
+
+module.exports = router;
