--- functions/routes/image.route.js
@@ -1,8 +1,0 @@
-const express = require("express");
-const imageController = require("../controllers/image.controller");
-
-const router = express.Router();
-
-router.get("/get", imageController.getImages);
-
-module.exports = router;
